ringbuf: Fix incorrect calculation available byte calculation
This fixes an incorrect check for whether there are
enough bytes available to service a read in `ringbuf_read`.

Fixing the check revealed an off-by-one in `ringbuf_free_space`
which is also resolved by this commit.

Due to the strict bounds checks on all pointer manipulations
in the ringbuf code, these bugs should have no security implications.